requirements.txt
numpy>=1.26
Pillow>=10.0
PyYAML>=6.0
torch>=2.2
diffusers>=0.30
transformers>=4.40
opencv-python>=4.9
scipy>=1.12
pandas>=2.2
streamlit>=1.34
accelerate>=0.30
safetensors>=0.4

